Add Dockerfile to build Knock with Docker
It is convenient to use a containerized build environment to avoid
the hassle with dependencies on the build machine. Furthermore, it
provides a consistent build setup for all users. And last but not
least, having Knock available as a (multi-arch) container in a
container registry makes it easy to run and eliminates the need for the
user to run `uname -ms` and perform manual download.

Signed-off-by: Joakim Roubert <joakim.roubert@gmail.com>